+++ next.config.js
@@ -1,6 +1,15 @@
 /** @type {import('next').NextConfig} */
 const nextConfig = {
-    output: 'export',
-  }
-  
-  module.exports = nextConfig
+  output: 'export',
+  async redirects() {
+    return [
+      {
+        source: '/',
+        destination: '/home-page',
+        permanent: true,
+      },
+    ];
+  },
+};
+
+module.exports = nextConfig;
