+++ src/app/cropping-resizing-view-settings.tsx
@@ -1,4 +1,4 @@
-import NumberInput from "./number-input";
+import NumberInput from "./home-page/number-input";
 
 export default function CroppingResizingViewSettings({
   viewSettings,
